moved and trimmed the bMarker container class.
